Increase the maximum size that is used when reading the ssh-host-key
Currently keys with 4096bits are ignored because in base64-encoding they are larger than the current limit of 2048bytes.

Note: as long as the default-key-size does not change the limit for writing the generated file 2 lines below does not need to be increased.

As per: http://tomatousb.org/forum/t-764466/support-for-ssh-host-keys-with-4096bits