inet6: prevent network storms caused by linux IPv6 routers
https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?h=v2.6.39&id=72b43d0898e97f588293b4a24b33c58c46633d81

Linux IPv6 forwards unicast packets, which are link layer multicasts...
The hole was present since day one. I was 100% this check is there,
but it is not. The problem shows itself, f.e. when Microsoft Network Load Balancer
runs on a network. This software resolves IPv6 unicast addresses to multicast MAC addresses.

Signed-off-by: Alexey Kuznetsov <kuznet@ms2.inr.ac.ru>
Signed-off-by: David S. Miller <davem@davemloft.net>