Use libcurl if available for DDNS
The existing HTTP request implementation can be slow. I've found that it
takes substantially longer when doing DDNS updates (with Cloudflare)
than it does if libcurl is used.
This commit adds code to use libcurl instead of the built-in
implementation if available, as determined by TCONFIG_BBT and
TCONFIG_NGINX, which are the two configuration settings that pull in
libcurl.