pppd: fixes from upstream
- pppd.h: Add missing headers

Signed-off-by: Rosen Penev <rosenp@gmail.com>
https://github.com/paulusmack/ppp/commit/03104baaa85c5537b704558848199c306b345190

- pppd: Don't free static string

Commit fcb076c ("Various fixes for errors found by coverity static
analysis (#109)", 2019-05-06) added statements to free the result
returned from get_first_ethernet().  However, the result of
get_first_ethernet() is not dynamically allocated, either on Linux
or Solaris.  Hence this removes the unnecessary (and dangerous)
free() statements.

Fixes: fcb076c ("Various fixes for errors found by coverity static analysis (#109)")
Reported-by: Florian Kohnhäuser <florian@kohnhaeuser.com>
Signed-off-by: Paul Mackerras <paulus@ozlabs.org>
https://github.com/paulusmack/ppp/commit/09f695f09e4ff6905388e15b106cd5669b05965e

- pppd: Limit memory accessed by string formats with max length specified

Currently, calls to [v]slprintf that have a string format (%s, %v,
%q) with a maximum length specified (e.g. %.20s) do a strlen() on
the string, and can therefore access memory beyond the maximum
length specified.  If the string is not null-terminated, this could
result in an out-of-bounds read.

This makes vslprintf use strnlen() in cases where a maximum length
has been specified, so that we don't access the string beyond the
maximum length that was given.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>
https://github.com/paulusmack/ppp/commit/b311e98b1d4775f7db36b81697ed8996809f3639

- pppd: Make sure word read from options file is null-terminated

If a word read from an options file was longer than MAXWORDLEN,
we could pass it to option_error() without null termination,
which could have lead to an out-of-bounds access in vslprintf.
Make sure word[] is null terminated in all cases.

Reported-by: Florian Kohnhäuser <florian@kohnhaeuser.com>
Signed-off-by: Paul Mackerras <paulus@ozlabs.org>
https://github.com/paulusmack/ppp/commit/c10c3c7cb3e779af5286430d5c9bb835a873ab32

- pppd: Avoid use of strnlen (and strlen) in vslprintf

Commit b311e98 ("pppd: Limit memory accessed by string formats with
max length specified") added calls to strnlen() in vslprintf().
Unfortunately, strnlen() is not provided in some standard C libraries.
This changes the code to avoid using strnlen().  Using the observation
that the number of characters we can use from the input string is
bounded by buflen, the number of bytes of output buffer available,
we can also avoid doing strlen() on a potentially long string.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>
https://github.com/paulusmack/ppp/commit/5d034034a61c37dc5d590753cb7a601d2ee2b871