rc: Drastically improve slow boot times caused by USB mass storage.
When USB storage is present upon booting a router, or when "hotplugging" it in while the router is already running, a kernel event is triggered which signals to hotplug2 to run "/sbin/hotplug block" which will mount a partition and also execute the "restart_nas_services" function. This function uses file locking to ensure that multiple hotplug events (mounting multiple partitions) run serially. It is virtually guarenteed that USB storage already present on boot through the kernel hotplugg event will attempt to run "restart_nas_services" even before the init process has gotten to the point of starting the nas services. This first hotplug event then takes the file lock, bringing the boot process to a halt as it cannot proceed if it doesn't have the first file lock. After this it is unclear to me how it works itself out to finish the boot process (and I'm too exasperated at this point to care) but you end up with a boot-to-internet time of ~5:15 when it really should be around ~1:15 (R7000 + 1 USB HDD w/ (1) swap partition and (3) ext4 partitions).

This commit ensures that hotplug events do no take this file lock from the init process by checking if the init process has started the nas services yet. The init process will create the notice file /var/notice/nas. If a hotplug event does not detect this notice then running "restart_nas_services" is skipped as the init process itself hasn't gotten to that point in the boot process and will handle starting the nas service when it gets to it.